+++ apps/web/src/feature/app/list/SelectAppServer.tsx
@@ -1,6 +1,6 @@
 import { SelectAppClient } from "@feature/app/list/SelectAppClient";
 import { findUser } from "@persistence/browser/server/find-user";
-import { fetchQuery } from "@persistence/database/fetchQuery";
+import { fetchQuery } from "@persistence/database/server/fetchQuery";
 import { GetApplicationsDocument } from "@v3/graphql/public/type";
 
 export const SelectAppServer = async () => {
